ecommerce-be/src/requirements.txt
Flask==3.0.3
gunicorn
psycopg2-binary
Flask-SQLAlchemy
Flask-JWT-Extended==4.6.0